PHPDoc for payment functions.php

+++ includes/payment-tracking-functions.php
@@ -4,7 +4,15 @@
 * functions used for tracking payments and earnings
 ***************************************************/
 
-// retrieve payments from the database
+/**
+ * Retrieve all payments from database
+ *
+ * @access      private
+ * @param       $offset INT The number to skip
+ * @param       $number INT The number to retrieve
+ * @return      array
+*/
+
 function rcp_get_payments( $offset = 0, $number = 20 ) {
 	global $wpdb, $rcp_payments_db_name;
 	if( $number > 0 ) {
@@ -21,7 +29,12 @@ function rcp_get_payments( $offset = 0, $number = 20 ) {
 }
 
 
-// returns the total number of payments recorded
+/**
+ * Retrieve the total number of payments in the database
+ *
+ * @access      private
+ * @return      int
+*/
 function rcp_count_payments() {
 	global $wpdb, $rcp_payments_db_name;
 	$count = get_transient( 'rcp_payments_count' );
@@ -32,6 +45,14 @@ function rcp_count_payments() {
 	return $count;
 }
 
+
+/**
+ * Retrieve total site earnings
+ *
+ * @access      private
+ * @return      float
+*/
+
 function rcp_get_earnings() {
 	global $wpdb, $rcp_payments_db_name;
 	$payments = get_transient( 'rcp_earnings' );
@@ -40,7 +61,7 @@ function rcp_get_earnings() {
 		// cache the payments query
 		set_transient( 'rcp_earnings', $payments, 10800 );
 	}
-	$total = 0;
+	$total = (float) 0.00;
 	if( $payments ) :
 		foreach( $payments as $payment ) :
 			$total = $total + $payment->amount;
@@ -49,36 +70,45 @@ function rcp_get_earnings() {
 	return $total;
 }
 
+
+/**
+ * Insert a payment into the database
+ *
+ * @access      private
+ * @param       $payment_data ARRAY The data to store
+ * @return      INT the ID of the new payment, or false if insertion fails
+*/
+
 function rcp_insert_payment( $payment_data = array() ) {
 	global $wpdb, $rcp_payments_db_name;
-		
+
 	$amount = $payment_data['amount'];
 	if( $payment_data['amount'] == '' )
 		$amount = $payment_data['amount2'];
-	
+
 	if( rcp_check_for_existing_payment( $payment_data['payment_type'], $payment_data['date'], $payment_data['subscription_key'] ) )
 		return;
 
-	$wpdb->insert( 
-		$rcp_payments_db_name, 
-		array( 
-			'subscription' 		=> $payment_data['subscription'], 
-			'date' 				=> $payment_data['date'], 
+	$wpdb->insert(
+		$rcp_payments_db_name,
+		array(
+			'subscription' 		=> $payment_data['subscription'],
+			'date' 				=> $payment_data['date'],
 			'amount' 			=> $amount,
 			'user_id' 			=> $payment_data['user_id'],
 			'payment_type' 		=> $payment_data['payment_type'],
 			'subscription_key' 	=> $payment_data['subscription_key']
-		), 
-		array( 
-			'%s', 
+		),
+		array(
+			'%s',
 			'%s',
 			'%s',
 			'%d',
 			'%s',
 			'%s'
-		) 
+		)
 	);
-	
+
 	// if insert was succesful, return the payment ID
 	if( $wpdb->insert_id ) {
 		// clear the payment caches
@@ -92,8 +122,19 @@ function rcp_insert_payment( $payment_data = array() ) {
 	return false;
 }
 
+
+/**
+ * Check if a payment already exists
+ *
+ * @access      private
+ * @param       $type string The type of payment (web_accept, subscr_payment, Credit Card, etc)
+ * @param       $date string/date The date of tpaen
+ * @param       $subscriptionkey string The subscription key the payment is connected to
+ * @return      bool
+*/
+
 function rcp_check_for_existing_payment( $type, $date, $subscription_key ) {
-	
+
 	global $wpdb, $rcp_payments_db_name;
 
 	if( $wpdb->get_results( $wpdb->prepare("SELECT id FROM " . $rcp_payments_db_name . " WHERE `date`='%d' AND `subscription_key`='%s' AND `payment_type`='%s';", $date, $subscription_key, $type ) ) )
